requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
attrs==22.2.0 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
-e ./coffee-freaks-api-client
fastapi==0.95.0
geonamescache==1.5.0
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
idna==3.4 ; python_version >= '3.5'
pydantic==1.10.7 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-multipart==0.0.6
rfc3986[idna2008]==1.5.0
setuptools==67.6.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.26.1 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
